+++ src/main/java/com/foodtech/blog/todoTask/service/TodoTaskApiService.java
@@ -1,44 +1,40 @@
 package com.foodtech.blog.todoTask.service;
 
-import com.foodtech.blog.base.api.request.SearchRequest;
+import com.foodtech.blog.auth.exceptions.AuthException;
+import com.foodtech.blog.auth.exceptions.NotAccessException;
+import com.foodtech.blog.auth.service.AuthService;
 import com.foodtech.blog.base.api.response.SearchResponse;
+import com.foodtech.blog.base.servise.CheckAccess;
 import com.foodtech.blog.todoTask.api.request.TodoTaskRequest;
 import com.foodtech.blog.todoTask.api.request.TodoTaskSearchRequest;
 import com.foodtech.blog.todoTask.mapping.TodoTaskMapping;
-import com.foodtech.blog.todoTask.api.response.TodoTaskResponse;
-import com.foodtech.blog.todoTask.exeception.TodoTaskExistException;
 import com.foodtech.blog.todoTask.exeception.TodoTaskNotExistException;
 import com.foodtech.blog.todoTask.model.TodoTaskDoc;
 import com.foodtech.blog.todoTask.repository.TodoTaskRepository;
-import com.foodtech.blog.user.exeception.UserNotExistException;
+import com.foodtech.blog.base.api.model.UserDoc;
 import com.foodtech.blog.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.util.DigestUtils;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class TodoTaskApiService {
+public class TodoTaskApiService extends CheckAccess<TodoTaskDoc> {
     private final TodoTaskRepository todoTaskRepository;
     private final MongoTemplate mongoTemplate;
     private final UserRepository userRepository;
+    private final AuthService authService;
 
-    public TodoTaskDoc create(TodoTaskRequest request) throws TodoTaskExistException, UserNotExistException {
-        if(userRepository.findById(request.getOwnerId()).isEmpty()) throw new UserNotExistException();
+    public TodoTaskDoc create(TodoTaskRequest request) throws  AuthException {
+        UserDoc userDoc =authService.currentUser();
 
-        TodoTaskDoc todoTaskDoc =TodoTaskMapping.getInstance().getRequest().convert(request);
+        TodoTaskDoc todoTaskDoc =TodoTaskMapping.getInstance().getRequest().convert(request,userDoc.getId());
         todoTaskRepository.save(todoTaskDoc);
         return  todoTaskDoc;
     }
@@ -48,10 +44,10 @@ public class TodoTaskApiService {
     }
     public SearchResponse<TodoTaskDoc> search(
              TodoTaskSearchRequest request
-    ){
-        if(request.getOwnerId() == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+    ) throws AuthException {
+        UserDoc userDoc = authService.currentUser();
 
-        Criteria criteria = Criteria.where("ownerId").is(request.getOwnerId());
+        Criteria criteria = Criteria.where("ownerId").is(userDoc.getId());
         if(request.getQuery() != null && request.getQuery()!=""){
             criteria = criteria.orOperator(
                     Criteria.where("title").regex(request.getQuery(), "i")
@@ -68,14 +64,14 @@ public class TodoTaskApiService {
         return SearchResponse.of(todoTaskDocs, count);
     }
 
-    public TodoTaskDoc update(TodoTaskRequest request) throws TodoTaskNotExistException {
+    public TodoTaskDoc update(TodoTaskRequest request) throws TodoTaskNotExistException, NotAccessException, AuthException {
         Optional<TodoTaskDoc> todoTaskDocOptional = todoTaskRepository.findById(request.getId());
         if(todoTaskDocOptional.isPresent() == false){
             throw new TodoTaskNotExistException();
         }
         TodoTaskDoc oldDoc = todoTaskDocOptional.get();
-
-        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequest().convert(request);
+        UserDoc userDoc = checkAccess(oldDoc);
+        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequest().convert(request, userDoc.getId());
 
         todoTaskDoc.setId(request.getId());
         todoTaskDoc.setOwnerId(oldDoc.getOwnerId());
@@ -84,7 +80,18 @@ public class TodoTaskApiService {
         return todoTaskDoc;
     }
 
-    public void delete(ObjectId id){
+    public void delete(ObjectId id) throws NotAccessException, AuthException {
+        checkAccess(todoTaskRepository.findById(id).get());
         todoTaskRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(TodoTaskDoc entity) {
+        return entity.getOwnerId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
